// src/types/propertyStorage/types.ts
export interface StringPropertyConfig extends BasePropertyConfig {
  controlType: 'string';
  placeholder: string;         // 输入占位符
}

// 数字控件配置
export interface NumberPropertyConfig extends BasePropertyConfig {
  controlType: 'number';
  min: number;                 // 最小值
  max: number;                 // 最大值
  step: number;                // 步长
  placeholder: string;         // 输入占位符
}

// 布尔控件配置
export interface BooleanPropertyConfig extends BasePropertyConfig {
  controlType: 'boolean';
}

// 选择控件配置
export interface SelectPropertyConfig extends BasePropertyConfig {
  controlType: 'select';
  options: Array<{ label: string; value: any }>; // 选项列表
  placeholder: string;         // 选择占位符
}

// 颜色控件配置
export interface ColorPropertyConfig extends BasePropertyConfig {
  controlType: 'color';
  presetColors: string[];      // 预设颜色列表
}

// 图片控件配置
export interface ImagePropertyConfig extends BasePropertyConfig {
  controlType: 'image';
  acceptTypes: string;         // 接受的文件类型
  placeholder: '输入背景图链接'
}

// 四值字段定义
export interface QuadValueField {
  key: string;                  // 字段名称
  label: string;                // 显示标签
  defaultValue: number;         // 默认值
  min: number;                 // 最小值
  max: number;                 // 最大值
  step: number;                // 步长
  width: string;               // 宽度
}

// 四值控件配置
export interface QuadValuePropertyConfig extends BasePropertyConfig {
  controlType: 'quadValue';
  fieldConfig: QuadValueField[]; // 四值字段配置
  layout: "grid" | "flex" | "stack"; // 布局类型
  gridCols: number;            // 网格列数
  groupLabel: string;          // 分组标签
}

// 变换控件配置
export interface TransformPropertyConfig extends BasePropertyConfig {
  controlType: 'transform';
}

// 滑块控件配置
export interface SliderPropertyConfig extends BasePropertyConfig {
  controlType: 'slider';
  min: number;                  // 最小值
  max: number;                  // 最大值
  step: number;                // 步长
  inputWidth: string;          // 输入框宽度
}

// 触发器控件配置
export interface TriggerPropertyConfig extends BasePropertyConfig {
  controlType: 'trigger';
  options: Array<{ label: string; value: string }>; // 触发器类型选项
}

// 重复次数控件配置
export interface RepeatCountPropertyConfig extends BasePropertyConfig {
  controlType: 'repeatCount';
  options: Array<{ label: string; value: string }>; // 重复次数选项
}

// 合并所有控件配置类型
export type propertyConfig =
  | StringPropertyConfig
  | NumberPropertyConfig
  | BooleanPropertyConfig
  | SelectPropertyConfig
  | ColorPropertyConfig
  | ImagePropertyConfig
  | QuadValuePropertyConfig
  | TransformPropertyConfig
  | SliderPropertyConfig
  | TriggerPropertyConfig
  | RepeatCountPropertyConfig;

// 属性存储接口
export interface propertyStorage {
  attributes: { [propertyName: string]: propertyConfig; };
  style: { [propertyName: string]: propertyConfig; };
}